Tags with descriptions test
Added a new test to determine whether it's best to embed tags with just the tag, or to include the description. It turns out that adding descriptions is better.

Also, which instructions are best varies a lot between just the tags and tags with descriptions. Not sure why this is yet.