Added organizations
Added data to test organization classification. Accompanying this, added new logic to check for "negative" results. These are results that are trying to intentionally game the search results classification. Whether or not we can actually filter these in this step (or if there needs to be some deboost logic somewhere else) is unclear